pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "alim"
version = "0.1.0"
description = "ALİM — Ağıllı Lokal İdarəetmə Mütəxəssisi for Azerbaijani Farmers"
authors = ["Rashad Jafarov <rashad.jafaroff@gmail.com>"]
readme = "README.md"
packages = [{include = "alim", from = "src"}]


[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.116.0"
uvicorn = {extras = ["standard"], version = "^0.35.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
httpx = "^0.27.0"
# LangGraph/LangChain Stack - Updated to 1.x versions (Jan 2026)
langgraph = "^1.0.7"
langgraph-api = "^0.7.8"
langgraph-cli = {extras = ["inmem"], version = "^0.4.12"}
langchain-core = "^1.2.0"
langchain = "^1.2.0"
langfuse = "^2.0.0"
sqlalchemy = "^2.0.0"
alembic = "^1.13.0"
redis = "^5.0.0"
faker = "^22.0.0"
jinja2 = "^3.1.0"
structlog = "^24.1.0"
protobuf = ">=6.31.1,<7.0.0"
aiosqlite = "^0.22.1"
asyncpg = "^0.31.0"
toml = "^0.10.2"
pre-commit = "^4.5.1"
detect-secrets = "^1.5.0"
# LangGraph Checkpointing - IMPORTANT: psycopg[binary] required on Windows
psycopg = {extras = ["binary"], version = "^3.1.0"}
langgraph-checkpoint-postgres = "^3.0.3"
langgraph-checkpoint-redis = "^0.3.2"
respx = "^0.22.0"
uv = "^0.9.26"

# ═══════════════════════════════════════════════════════════════════════════
# 🖥️ Chainlit UI Dependencies (Unified - No Separate Venv)
# ═══════════════════════════════════════════════════════════════════════════
chainlit = "^2.9.6"
python-dotenv = "^1.0.0"
langchain-ollama = "^1.0.0"
langchain-anthropic = "^1.3.0"
langchain-openai = "^1.1.0"
langchain-mcp-adapters = "^0.2.0"
plotly = ">=5.18.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
ruff = "^0.1.0"
mypy = "^1.8.0"

# ═══════════════════════════════════════════════════════════════════════════
# 🚀 Poetry Scripts - Run commands without manually activating venv
# ═══════════════════════════════════════════════════════════════════════════
# Usage:
#   poetry run dev        # Start FastAPI server with auto-reload
#   poetry run migrate    # Run database migrations
#   poetry run seed       # Seed database with test data
#   poetry run test       # Run pytest
#   poetry run lint       # Run ruff linter
#
# Or activate shell once and use commands directly:
#   poetry shell
#   uvicorn ALİM.api.main:app --reload
#   alembic upgrade head
# ═══════════════════════════════════════════════════════════════════════════
[tool.poetry.scripts]
dev = "scripts.dev:main"
migrate = "scripts.migrate:main"
seed = "scripts.seed:main"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
markers = [
    "integration: Integration tests requiring external services",
    "ollama: Tests requiring local Ollama server",
    "cloud: Tests requiring cloud API keys (Groq)",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
